Add Policies API to patch panel
This adds a /policies child to each interface, with the intended use
that a module that implements a rx/tx handler but takes no forwarding
action can be installed in the packet path of a different module that
does take forwarding decisions.

Transition GBP to be a module that implements the above interface.

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>