Preserve skb metadata across netdev boundary
A chain of ebpf->linuxbridge->ebpf can lose the metadata of a packet, in
cases where the bridge enqueues the packet (e.g. fq_codel). So, instead,
keep the metadata of the packet in a hash that is shared with a kprobe
bpf program. The metadata is allocated when the skb leaves the bpf chain
and enters a netdev, and is recovered when received again from a netdev.
The metadata is freed when kfree_skbmem is called in the kernel.

This is not an ideal solution by any means, and should be replaced when
one of the following alternatives is available:
 more reliable metadata in the skb
 an LRU style hash
 we can gaurantee that a packet will not be rescheduled

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>