Implement packet out action
This commit adds the possibility for the controller to send a packet to
an IOModule.
The packet can be injected using to different strategies, the "INGRESS"
strategy simulates the packet arriving any IOModule iface, while the
"EGRESS" strategy sends the packet through a port of the IOModule.

The INGRESS implementation is straightforward as the destination module
is known. In the EGRESS implementation it is necessary to perform a look
up on the graph in order to get the next IOModule