implement packet broadcast action
Connections between IOModules are implemented using eBPF tailcalls in
hover, due to the characteristics of the tailcalls it is impossible
for an IOModule to send the same packet to multiple IOModules.

This commit adds a feature in the local controller that broadcasts a
packet upon an IOModule request. When an IOModule wants to broadcast a
packet it should do

pkt_controller(skb, md, PKT_BROADCAST);
return RX_CONTROLLER;

Notes:
- Multicast is currently not supported
- A mechanism to avoid concurrency problems when sending packet to the
  IOModules is missing
- A better error handling has to be implemented